2019.06.01 [FIX] db error 수정

+++ app/src/main/java/konkukSW/MP2019/roadline/Data/Dataclass/Plan.kt
@@ -3,5 +3,5 @@ package konkukSW.MP2019.roadline.Data.Dataclass
 import java.io.Serializable
 
 data class Plan (var listID:String, var dayNum:Int, var id: String, var name:String,
-                 var locationX:Float, var locationY:Float, var time:String, var memo:String, var viewType:Int) :Serializable {
+                 var locationX:Double, var locationY:Double, var time:String, var memo:String, var viewType:Int) :Serializable {
 }
